Changes:
- Update README.md
- Update .gitignore
- Change border radius for all views
- Remove scrollbar.xml file
- Add activity_main_scene.xml file
- Update activity_main.xml and fragment_bar.xml files
- Update strings.xml
- Refactor Activities
- Refactor Fragments
- Move Activity classes to activities package
- Add PreferenceManager for working with SharedPreferences
- Refactor & Update layout files
- Add dimens.xml
- Add EmptyFragment class
- Add new animations and update old animations
- Refactor Password database classes
- Refactor utils package classes
- Rename drawable files
- Refactor xml resource files
- Update app-release.xml

Solves issues:
- Solve #42
- Solve #43
- Solve #44
- Solve #46
- Solve #47
- Solve #48

+++ app/src/main/java/com/merive/securely/utils/Hex.java
@@ -6,18 +6,25 @@ import androidx.annotation.RequiresApi;
 
 public class Hex {
 
+    /**
+     * @param value String that will be encrypted
+     * @return Return encrypted string of hex codes
+     */
     @RequiresApi(api = Build.VERSION_CODES.O)
     public static String encrypt(String value) {
         String[] encryptedSymbols = new String[value.length()];
         for (int i = 0; i < value.length(); i++)
-            encryptedSymbols[i] = Integer.toHexString((int) value.toCharArray()[i]);
+            encryptedSymbols[i] = Integer.toHexString(value.toCharArray()[i]);
         return String.join(" ", encryptedSymbols);
     }
 
+    /**
+     * @param value Encrypted string of hex codes
+     * @return Return decrypted string
+     */
     public static String decrypt(String value) {
         StringBuilder decryptedString = new StringBuilder();
-        for (String s : value.split(" "))
-            decryptedString.append((char) Integer.parseInt(s, 16));
+        for (String s : value.split(" ")) decryptedString.append((char) Integer.parseInt(s, 16));
         return decryptedString.toString();
     }
 }
